+++ @xen-orchestra/lite/src/views/host/HostRootView.vue
@@ -1,5 +1,9 @@
 <template>
-  <RouterView />
+  <ObjectNotFoundWrapper object-type="host">
+    <RouterView />
+  </ObjectNotFoundWrapper>
 </template>
 
-<script lang="ts" setup></script>
+<script lang="ts" setup>
+import ObjectNotFoundWrapper from "@/components/ObjectNotFoundWrapper.vue";
+</script>